# Conflicts:
#	Source/LinqToDB/Expressions/InternalExtensions.cs
#	Source/LinqToDB/Linq/Builder/AggregationBuilder.cs
#	Source/LinqToDB/Linq/Builder/ExpressionBuilder.SqlBuilder.cs
#	Source/LinqToDB/Linq/Builder/MethodChainBuilder.cs

+++ Data/Create Scripts/Sybase.sql
@@ -271,8 +271,8 @@ SELECT * FROM Person
 GO
 
 CREATE TABLE KeepIdentityTest (
-	ID    NUMERIC(12, 0) IDENTITY,
-	Value INT            NULL
+	ID    INT IDENTITY,
+	Value INT NULL
 )
 GO
